+++ examples/actor_animation/GlutWindow.cc
@@ -29,6 +29,7 @@
   #include <GL/glx.h>
 #endif
 
+#include <iomanip>
 #include <mutex>
 #include <map>
 #include <string>
@@ -59,12 +60,17 @@ ir::CameraPtr g_camera;
 ir::CameraPtr g_currCamera;
 unsigned int g_cameraIndex = 0;
 ir::ImagePtr g_image;
-ir::MeshPtr g_mesh;
+std::vector<ir::VisualPtr> g_visuals;
+std::vector<ir::VisualPtr> g_allVisuals;
 ic::SkeletonPtr g_skel;
 ic::SkeletonAnimation *g_skelAnim;
-unsigned int g_animIdx = 1;
+unsigned int g_animIdx = 0;
+bool g_updateAll = false;
+bool g_manualBoneUpdate = false;
+bool g_actorUpdateDirty = true;
 std::chrono::steady_clock::duration g_time{0};
 std::chrono::steady_clock::time_point g_startTime;
+std::chrono::steady_clock::time_point prevUpdateTime;
 
 bool g_initContext = false;
 
@@ -239,42 +245,135 @@ void handleMouse()
 //////////////////////////////////////////////////
 void updatePose(double _time)
 {
-  std::map<std::string, ignition::math::Matrix4d> animFrames;
-  animFrames = g_skelAnim->PoseAt(_time, true);
-
-  std::map<std::string, ignition::math::Matrix4d> skinFrames;
-
-  for (auto pair : animFrames)
+  // manually update the bone pose
+  for (auto &v : g_visuals)
   {
-    std::string animName = pair.first;
-    auto animTf = pair.second;
+    std::map<std::string, ignition::math::Matrix4d> animFrames;
+    animFrames = g_skelAnim->PoseAt(_time, true);
+    std::map<std::string, ignition::math::Matrix4d> skinFrames;
+    for (auto pair : animFrames)
+    {
+      std::string animNodeName = pair.first;
+      auto animTf = pair.second;
 
-    std::string skinName = g_skel->NodeNameAnimToSkin(g_animIdx, animName);
-    ignition::math::Matrix4d skinTf =
-            g_skel->AlignTranslation(g_animIdx, animName)
-            * animTf * g_skel->AlignRotation(g_animIdx, animName);
+      std::string skinName =
+          g_skel->NodeNameAnimToSkin(g_animIdx, animNodeName);
+      ignition::math::Matrix4d skinTf =
+              g_skel->AlignTranslation(g_animIdx, animNodeName)
+              * animTf * g_skel->AlignRotation(g_animIdx, animNodeName);
 
-    skinFrames[skinName] = skinTf;
+      skinFrames[skinName] = skinTf;
+    }
+
+    ir::MeshPtr mesh =
+        std::dynamic_pointer_cast<ir::Mesh>(v->GeometryByIndex(0));
+    mesh->SetSkeletonLocalTransforms(skinFrames);
   }
+}
 
-  g_mesh->SetSkeletonLocalTransforms(skinFrames);
+//////////////////////////////////////////////////
+void updateTime(double _time)
+{
+  // set time to advance animation
+  for (auto &v : g_visuals)
+  {
+    ir::MeshPtr mesh =
+        std::dynamic_pointer_cast<ir::Mesh>(v->GeometryByIndex(0));
+    mesh->UpdateSkeletonAnimation(
+        std::chrono::duration_cast<std::chrono::steady_clock::duration>(
+        std::chrono::duration<double>(_time)));
+  }
 }
 
 //////////////////////////////////////////////////
 void updateActor()
 {
+  g_time = std::chrono::steady_clock::now() - g_startTime;
   auto seconds =
       std::chrono::duration_cast<std::chrono::milliseconds>(g_time).count() /
       1000.0;
-  updatePose(seconds);
-  if (seconds >= g_skelAnim->Length())
+
+  g_skelAnim = g_skel->Animation(g_animIdx);
+
+  // change detected due to key press
+  if (g_actorUpdateDirty)
   {
-    g_time = std::chrono::steady_clock::duration();
+    // disable all auto animations
+    for (auto &v : g_allVisuals)
+    {
+      ir::MeshPtr mesh =
+          std::dynamic_pointer_cast<ir::Mesh>(v->GeometryByIndex(0));
+      for (unsigned int i = 0; i < g_skel->AnimationCount(); ++i)
+      {
+        auto anim = g_skel->Animation(i);
+
+        // disable all
+        mesh->SetSkeletonAnimationEnabled(anim->Name(), false, false, 0.0);
+      }
+      v->SetVisible(false);
+    }
+
+    // set visuals to be updated
+    g_visuals.clear();
+    if (g_updateAll)
+    {
+      std::copy(g_allVisuals.begin(), g_allVisuals.end(),
+          std::back_inserter(g_visuals));
+    }
+    else
+    {
+      unsigned int idx = static_cast<unsigned int>(
+          sqrt(g_allVisuals.size()) * 0.5);
+      g_visuals.push_back(g_allVisuals[idx]);
+    }
+
+    // enabled selected animation
+    for (auto &v : g_visuals)
+    {
+      v->SetVisible(true);
+      if (!g_manualBoneUpdate)
+      {
+        ir::MeshPtr mesh =
+            std::dynamic_pointer_cast<ir::Mesh>(v->GeometryByIndex(0));
+        mesh->SetSkeletonAnimationEnabled(g_skelAnim->Name(), true, true, 1.0);
+      }
+    }
+
+
+    g_actorUpdateDirty = false;
   }
+
+  // manually update skeleton bone pose
+  if (g_manualBoneUpdate)
+  {
+    updatePose(fmod(seconds, g_skelAnim->Length()));
+  }
+  // advance time for built in animations
   else
   {
-    g_time = std::chrono::steady_clock::now() - g_startTime;
+    updateTime(seconds);
   }
+}
+
+//////////////////////////////////////////////////
+void drawText(int _x, int _y, const std::string &_text)
+{
+  glMatrixMode(GL_PROJECTION);
+  glPushMatrix();
+  glLoadIdentity();
+  gluOrtho2D(0.0, imgw, 0.0, imgh);
+  glMatrixMode(GL_MODELVIEW);
+  glPushMatrix();
+  glLoadIdentity();
+  glColor3f(1.0f, 1.0f, 1.0f);
+  glRasterPos2i(_x, _y);
+  void *font = GLUT_BITMAP_9_BY_15;
+  for (auto c : _text)
+    glutBitmapCharacter(font, c);
+  glMatrixMode(GL_MODELVIEW);
+  glPopMatrix();
+  glMatrixMode(GL_PROJECTION);
+  glPopMatrix();
 }
 
 //////////////////////////////////////////////////
@@ -308,9 +407,20 @@ void displayCB()
   glRasterPos2f(-1, 1);
   glDrawPixels(imgw, imgh, GL_RGB, GL_UNSIGNED_BYTE, data);
 
-  glutSwapBuffers();
-
   updateActor();
+
+  // draw FPS
+  auto now = std::chrono::steady_clock::now();
+  double t = std::chrono::duration_cast<std::chrono::milliseconds>(
+      now - prevUpdateTime).count() / 1000.0;
+  prevUpdateTime = now;
+  std::string manual = g_manualBoneUpdate ? "true" : "false";
+  std::stringstream text;
+  text << std::fixed << std::setw(5) << std::setprecision(4) << (1.0/t);
+  text << std::setw(30) << "Manual skeleton update: " << manual;
+  drawText(10, 10, text.str());
+
+  glutSwapBuffers();
 }
 
 //////////////////////////////////////////////////
@@ -329,6 +439,26 @@ void keyboardCB(unsigned char _key, int, int)
   else if (_key == KEY_TAB)
   {
     g_cameraIndex = (g_cameraIndex + 1) % g_cameras.size();
+  }
+  else if (_key == 'a')
+  {
+    g_animIdx = (g_animIdx + 1) % 2;
+    auto anim = g_skel->Animation(g_animIdx);
+    g_actorUpdateDirty = true;
+    std::cout << "Playing animation: " << anim->Name() << std::endl;
+  }
+  else if (_key == 'm')
+  {
+    g_manualBoneUpdate = !g_manualBoneUpdate;
+    g_actorUpdateDirty = true;
+    std::cout << "Manual skeleton bone update: " << g_manualBoneUpdate
+              << std::endl;
+  }
+  else if (_key == 't')
+  {
+    g_updateAll = !g_updateAll;
+    g_actorUpdateDirty = true;
+    std::cout << "Update all meshes " << g_updateAll << std::endl;
   }
 }
 
@@ -366,11 +496,6 @@ void initAnimation()
     std::cerr << "Failed to load animation." << std::endl;
     return;
   }
-  const std::string RESOURCE_PATH =
-      ic::joinPaths(std::string(PROJECT_BINARY_PATH), "media");
-  std::string bvhFile = ic::joinPaths(RESOURCE_PATH, "cmu-13_26.bvh");
-  double scale = 0.055;
-  g_skel->AddBvhAnimation(bvhFile, scale);
 
   g_skelAnim = g_skel->Animation(g_animIdx);
 
@@ -380,15 +505,19 @@ void initAnimation()
 //////////////////////////////////////////////////
 void printUsage()
 {
-  std::cout << "===============================" << std::endl;
-  std::cout << "  TAB - Switch render engines  " << std::endl;
-  std::cout << "  ESC - Exit                   " << std::endl;
-  std::cout << "===============================" << std::endl;
+  std::cout << "========================================" << std::endl;
+  std::cout << "  TAB - Switch render engines           " << std::endl;
+  std::cout << "  ESC - Exit                            " << std::endl;
+  std::cout << "  A   - Switch animation                " << std::endl;
+  std::cout << "  M   - Toggle manual skeleton update   " << std::endl;
+  std::cout << "  T   - Toggle animated mesh count      " << std::endl;
+  std::cout << "========================================" << std::endl;
 }
 
 //////////////////////////////////////////////////
-void run(std::vector<ir::CameraPtr>_cameras,
-            ir::MeshPtr _mesh, ic::SkeletonPtr _skel)
+void run(std::vector<ir::CameraPtr> _cameras,
+         const std::vector<ir::VisualPtr> &_visuals,
+         ic::SkeletonPtr _skel)
 {
   if (_cameras.empty())
   {
@@ -406,7 +535,7 @@ void run(std::vector<ir::CameraPtr>_cameras,
 #endif
 
   g_cameras = _cameras;
-  g_mesh = _mesh;
+  g_allVisuals = _visuals;
   g_skel = _skel;
 
   initCamera(_cameras[0]);