+++ examples/gazebo_scene_viewer/SceneManager.hh
@@ -17,6 +17,15 @@
 #ifndef IGNITION_RENDERING_EXAMPLES_SCENEMANAGER_HH
 #define IGNITION_RENDERING_EXAMPLES_SCENEMANAGER_HH
 
+#if __cplusplus > 201703L
+    #include <filesystem>
+    using fs = std::filesystem;
+#else
+    #include <experimental/filesystem>
+    using namespace std;
+    namespace fs = std::experimental::filesystem::v1;
+#endif
+
 #include <string>
 #include <ignition/common/SingletonT.hh>
 #include "ignition/rendering/RenderTypes.hh"