+++ include/ignition/rendering/Material.hh
@@ -165,11 +165,11 @@ namespace ignition
       /// \param[in] _alpha Set the alpha threshold value
       /// \param[in] _twoSided Enable two sided rendering
       public: virtual void SetAlphaFromTexture(bool _enabled,
-                  double _alpha=0.5, bool _twoSided=true) = 0;
+                  double _alpha = 0.5, bool _twoSided = true) = 0;
 
       /// \brief Get the enable alpha from textures
       /// \return The enable alpha value
-      public: virtual bool AlphaEnabled() const = 0;
+      public: virtual bool TextureAlphaEnabled() const = 0;
 
       /// \brief Get the alpha threshold
       /// \return The alpha threshold value