+++ src/Mesh_TEST.cc
@@ -19,6 +19,9 @@
 #include <string>
 
 #include <ignition/common/Console.hh>
+#include <ignition/common/MeshManager.hh>
+#include <ignition/common/Skeleton.hh>
+#include <ignition/common/SkeletonAnimation.hh>
 
 #include "test_config.h"  // NOLINT(build/include)
 #include "ignition/rendering/Camera.hh"
@@ -35,6 +38,13 @@ class MeshTest : public testing::Test,
 {
   /// \brief Test mesh and submesh basic API
   public: void MeshSubMesh(const std::string &_renderEngine);
+
+  /// \brief Test mesh skeleton animation API
+  public: void MeshSkeletonAnimation(const std::string &_renderEngine);
+
+  public: const std::string TEST_MEDIA_PATH =
+        common::joinPaths(std::string(PROJECT_SOURCE_PATH),
+        "test", "media", "meshes");
 };
 
 /////////////////////////////////////////////////
@@ -94,6 +104,73 @@ void MeshTest::MeshSubMesh(const std::string &_renderEngine)
 TEST_P(MeshTest, MeshSubMesh)
 {
   MeshSubMesh(GetParam());
+}
+
+/////////////////////////////////////////////////
+void MeshTest::MeshSkeletonAnimation(const std::string &_renderEngine)
+{
+  RenderEngine *engine = rendering::engine(_renderEngine);
+  if (!engine)
+  {
+    igndbg << "Engine '" << _renderEngine
+              << "' is not supported" << std::endl;
+    return;
+  }
+
+  ScenePtr scene = engine->CreateScene("scene");
+  ASSERT_TRUE(scene != nullptr);
+
+  // test box mesh with no skeleton animation
+  MeshDescriptor boxDescriptor("unit_box");
+  MeshPtr boxMesh = scene->CreateMesh(boxDescriptor);
+  ASSERT_TRUE(boxMesh != nullptr);
+
+  EXPECT_FALSE(boxMesh->HasSkeleton());
+  EXPECT_NO_THROW(boxMesh->SetSkeletonAnimationEnabled("invalid", false));
+  EXPECT_FALSE(boxMesh->SkeletonAnimationEnabled("invalid"));
+  EXPECT_TRUE(boxMesh->SkeletonLocalTransforms().empty());
+
+  // test mesh with skeleton animation
+  MeshDescriptor descriptor;
+  descriptor.meshName = common::joinPaths(TEST_MEDIA_PATH, "walk.dae");
+  common::MeshManager *meshManager = common::MeshManager::Instance();
+  descriptor.mesh = meshManager->Load(descriptor.meshName);
+  MeshPtr mesh = scene->CreateMesh(descriptor);
+
+  EXPECT_TRUE(mesh->HasSkeleton());
+  EXPECT_FALSE(mesh->SkeletonLocalTransforms().empty());
+
+  auto skel = descriptor.mesh->MeshSkeleton();
+  ASSERT_EQ(1u, skel->AnimationCount());
+
+  std::string animName = skel->Animation(0u)->Name();
+  EXPECT_FALSE(mesh->SkeletonAnimationEnabled(animName));
+
+  mesh->SetSkeletonAnimationEnabled(animName, true);
+  EXPECT_TRUE(mesh->SkeletonAnimationEnabled(animName));
+
+  EXPECT_NO_THROW(mesh->UpdateSkeletonAnimation(
+      std::chrono::duration_cast<std::chrono::steady_clock::duration>(
+      std::chrono::duration<double>(0.0))));
+  EXPECT_NO_THROW(mesh->UpdateSkeletonAnimation(
+      std::chrono::duration_cast<std::chrono::steady_clock::duration>(
+      std::chrono::duration<double>(1.0))));
+  EXPECT_NO_THROW(mesh->UpdateSkeletonAnimation(
+      std::chrono::duration_cast<std::chrono::steady_clock::duration>(
+      std::chrono::duration<double>(-11.0))));
+  EXPECT_NO_THROW(mesh->UpdateSkeletonAnimation(
+      std::chrono::duration_cast<std::chrono::steady_clock::duration>(
+      std::chrono::duration<double>(1234.5))));
+
+  // Clean up
+  engine->DestroyScene(scene);
+  rendering::unloadEngine(engine->Name());
+}
+
+/////////////////////////////////////////////////
+TEST_P(MeshTest, MeshSkeletonAnimation)
+{
+  MeshSkeletonAnimation(GetParam());
 }
 
 INSTANTIATE_TEST_CASE_P(Mesh, MeshTest,